ref: convert float64 to int64 instead of int to prevent precision loss and ensure consistent behavior across platforms

+++ pkg/razorpay/server_test.go
@@ -111,7 +111,7 @@ func Test_RequiredInt(t *testing.T) {
 		name        string
 		params      map[string]interface{}
 		paramName   string
-		expected    int
+		expected    int64
 		expectError bool
 	}{
 		{
@@ -157,7 +157,7 @@ func Test_OptionalInt(t *testing.T) {
 		name        string
 		params      map[string]interface{}
 		paramName   string
-		expected    int
+		expected    int64
 		expectError bool
 	}{
 		{
